test: move _logShadowDebug coverage into percy.test.js
Inline the unit test next to the existing closed-shadow snapshot test
instead of resurrecting a dedicated packages/core/test/unit/page.test.js
file. Same coverage, no new test file.
